% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{DLMextra-OM}
\alias{DLMextra-OM}
\alias{DLMextra}
\alias{DLMextra-package}
\alias{Albacore_IO_IOTC}
\alias{Almaco_Jack_GOM_NOAA}
\alias{Arrowtooth_Flounder_BC_DFO}
\alias{Atlantic_Herring_4TFallSpawn_DFO}
\alias{Atlantic_Herring_4TSpringSpawn_DFO}
\alias{Banded_Morwong_Tasmania}
\alias{Barred_Sand_Bass_CA_CDFW}
\alias{Bigeye_ICCAT}
\alias{Blue_Shark_IO_IOTC}
\alias{Blue_Shark2_IO_IOTC}
\alias{Bluefin_Tuna_EAtl_ICCAT}
\alias{Bluefin_Tuna_WAtl_ICCAT}
\alias{California_Halibut_CA_CDFW}
\alias{Canary_Rockfish_BC_DFO}
\alias{Capelin_GSL_DFO}
\alias{Cod_4T4Vn_DFO}
\alias{Cod_4X5Y_DFO}
\alias{Cod_5ZJM_DFO}
\alias{Gatuzo_Argentina_INIDEP}
\alias{Grey_Seal_5ZJM_DFO}
\alias{Haddock_4X5Y_DFO}
\alias{Haddock_Faroe_Islands}
\alias{Halibut_Atl_DFO}
\alias{Hogfish_PR_NOAA}
\alias{Jonah_Crab_LFA34_DFO}
\alias{Lane_Snapper_GOM_NOAA}
\alias{Lesser_Amberjack_GOM_NOAA}
\alias{Pacific_Hake_USCan}
\alias{Pacific_Ocean_Perch_QC_BC_DFO}
\alias{Palo_Argentina_INIDEP}
\alias{Queen_Triggerfish_STT_NOAA}
\alias{Red_Drum_GOM_NOAA}
\alias{Red_Sea_Urchin_CA_CDFW}
\alias{Redbanded_Rockfish_BC_DFO}
\alias{Rougheye_Rockfish_BC_DFO}
\alias{Shortspine_Thornyhead_BC_DFO}
\alias{Skipjack_IO_IOTC}
\alias{Snowy_Grouper_GOM_NOAA}
\alias{Speckled_Hind_GOM_NOAA}
\alias{Spiny_Lobster_STX_NOAA}
\alias{Stones_Sheep_BC}
\alias{Stoplight_Parrotfish_STX_NOAA}
\alias{Swordfish_ICCAT}
\alias{Swordfish_IO_IOTC}
\alias{Swordfish_NAtl_DFO}
\alias{Tiger_flathead_SEAus_AFMA}
\alias{Toothfish_HeardIsland_CCAMLR}
\alias{Vermillion_Snapper_GOM_NOAA}
\alias{Warty_Sea_Cucumber_CA_CDFW}
\alias{Yelloweye_Rockfish_BC_DFO}
\alias{Yellowfin_Tuna_IO_IOTC}
\alias{Yellowfin_Tuna2_IOTC}
\alias{Yellowmouth_Grouper_GOM_NOAA}
\alias{Yellowtail_Snapper_PR_NOAA}
\alias{testMSE}
\title{Operating models and MSE output for DLMtool}
\format{An object of class \code{OM} of length 1.}
\usage{
Albacore_IO_IOTC

Almaco_Jack_GOM_NOAA

Arrowtooth_Flounder_BC_DFO

Atlantic_Herring_4TFallSpawn_DFO

Atlantic_Herring_4TSpringSpawn_DFO

Banded_Morwong_Tasmania

Barred_Sand_Bass_CA_CDFW

Bigeye_ICCAT

Blue_Shark_IO_IOTC

Blue_Shark2_IO_IOTC

Bluefin_Tuna_EAtl_ICCAT

Bluefin_Tuna_WAtl_ICCAT

California_Halibut_CA_CDFW

Canary_Rockfish_BC_DFO

Capelin_GSL_DFO

Cod_4T4Vn_DFO

Cod_4X5Y_DFO

Cod_5ZJM_DFO

Gatuzo_Argentina_INIDEP

Grey_Seal_5ZJM_DFO

Haddock_4X5Y_DFO

Haddock_Faroe_Islands

Halibut_Atl_DFO

Hogfish_PR_NOAA

Jonah_Crab_LFA34_DFO

Lane_Snapper_GOM_NOAA

Lesser_Amberjack_GOM_NOAA

Pacific_Hake_USCan

Pacific_Ocean_Perch_QC_BC_DFO

Palo_Argentina_INIDEP

Queen_Triggerfish_STT_NOAA

Red_Drum_GOM_NOAA

Red_Sea_Urchin_CA_CDFW

Redbanded_Rockfish_BC_DFO

Rougheye_Rockfish_BC_DFO

Shortspine_Thornyhead_BC_DFO

Skipjack_IO_IOTC

Snowy_Grouper_GOM_NOAA

Speckled_Hind_GOM_NOAA

Spiny_Lobster_STX_NOAA

Stones_Sheep_BC

Stoplight_Parrotfish_STX_NOAA

Swordfish_ICCAT

Swordfish_IO_IOTC

Swordfish_NAtl_DFO

Tiger_flathead_SEAus_AFMA

Toothfish_HeardIsland_CCAMLR

Vermillion_Snapper_GOM_NOAA

Warty_Sea_Cucumber_CA_CDFW

Yelloweye_Rockfish_BC_DFO

Yellowfin_Tuna_IO_IOTC

Yellowfin_Tuna2_IOTC

Yellowmouth_Grouper_GOM_NOAA

Yellowtail_Snapper_PR_NOAA

testMSE
}
\description{
A suite of operating models (class OM) from the DLMtool website (\url{http://www.datalimitedtoolkit.org/fishery_library}) and a sample MSE (\code{testMSE})output object.
}
\examples{
library(DLMextra)
myMSE <- DLMtool::runMSE(Shortspine_Thornyhead_BC_DFO)
plot(Shortspine_Thornyhead_BC_DFO)
}
\references{
\url{http://www.datalimitedtoolkit.org/fishery_library}
}
\keyword{datasets}
