% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmxtra.R
\name{dlmFilter2}
\alias{dlmFilter2}
\title{Filter a dlm}
\usage{
dlmFilter2(y, object, simplify = FALSE)
}
\arguments{
\item{y}{data}

\item{object}{a dlm object}

\item{simplify}{FALSE}
}
\value{
Same as when calling dlmFilter

Same as when calling dlmFilter
}
\description{
A wrapper for dlmFilter

Not many details for now

A wrapper for dlmFilter but restricted to using R code
    in the filter.

Not many details for now
}
\author{
Søren Højsgaard

Søren Højsgaard
}
