% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmxtra.R
\name{dlmxFilter}
\alias{dlmxFilter}
\title{Filter a dlm}
\usage{
dlmxFilter(y, object, simplify = FALSE)
}
\arguments{
\item{y}{The data. ‘y’ can be a vector, a matrix, a univariate or
multivariate time series.}

\item{object}{An object of class ‘dlm’, or a list with components ‘m0’,
‘C0’, ‘FF’, ‘V’, ‘GG’, ‘W’, and optionally ‘JFF’, ‘JV’,
‘JGG’, ‘JW’, and ‘X’, defining the model and the parameters
of the prior distribution.}

\item{simplify}{FALSE}
}
\value{
Same as when calling dlmFilter

Same as when calling dlmFilter
}
\description{
A wrapper for dlmFilter

Not many details for now

A wrapper for dlmFilter but restricted to using R code
    in the filter.
}
\author{
Søren Højsgaard

Søren Højsgaard
}
