% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extract-dlm}
\alias{extract-dlm}
\alias{getdlm}
\alias{getdlm.dlm}
\alias{FFeval}
\alias{GGeval}
\alias{Veval}
\alias{Weval}
\alias{extract-dlmFiltered}
\alias{getdlm.dlmFiltered}
\alias{mm}
\alias{mm.dlmFiltered}
\alias{aa}
\alias{CC}
\alias{RR}
\title{Extractor for dlm and dlmFiltered objects}
\usage{
getdlm(object, name, select = TRUE)

\method{getdlm}{dlm}(object, name, select)

FFeval(object, select)

GGeval(object, select)

Veval(object, select)

Weval(object, select)

getdlm(object, name, select = TRUE)

\method{getdlm}{dlmFiltered}(object, name, select = TRUE)

mm(object, select = TRUE)

\method{mm}{dlmFiltered}(object, select = TRUE)

aa(object, select = TRUE)

CC(object, select = TRUE)

RR(object, select = TRUE)
}
\arguments{
\item{object}{A dlmFiltered object.}

\item{name}{Which name in the object to select.}

\item{select}{Which element to select.}
}
\value{
The relevant result

The relevant result
}
\description{
Extractor for dlm and dlmFiltered objects

Extractor for dlmFiltered objects
}
\author{
Søren Højsgaard

Søren Højsgaard
}
