% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmXtra.R
\name{extract-dlmFiltered}
\alias{extract-dlmFiltered}
\alias{mm}
\alias{mm.dlmFiltered}
\alias{aa}
\alias{aa.dlmFiltered}
\alias{CC}
\alias{CC.dlmFiltered}
\alias{RR}
\alias{RR.dlmFiltered}
\title{Extractor for dlmFiltered objects}
\usage{
mm(object, select = TRUE)

\method{mm}{dlmFiltered}(object, select = TRUE)

aa(object, select = TRUE)

\method{aa}{dlmFiltered}(object, select = TRUE)

CC(object, select = TRUE)

\method{CC}{dlmFiltered}(object, select = TRUE)

RR(object, select = TRUE)

\method{RR}{dlmFiltered}(object, select = TRUE)
}
\arguments{
\item{object}{A dlmFiltered object}

\item{select}{Which element to select}
}
\value{
The relevant result
}
\description{
Extractor for dlmFiltered objects
}
\author{
Søren Højsgaard
}
