% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmXtra.R
\name{extract-filtered}
\alias{extract-filtered}
\alias{mm}
\alias{mm.dlmFiltered}
\alias{aa}
\alias{aa.dlmFiltered}
\alias{CC}
\alias{CC.dlmFiltered}
\alias{RR}
\alias{RR.dlmFiltered}
\title{Extractor for dlmFiltered objects}
\usage{
mm(x, select = TRUE)

\method{mm}{dlmFiltered}(object, select = TRUE)

aa(x, select = TRUE)

\method{aa}{dlmFiltered}(object, select = TRUE)

CC(x, select = TRUE)

\method{CC}{dlmFiltered}(object, select = TRUE)

RR(x, select = TRUE)

\method{RR}{dlmFiltered}(object, select = TRUE)
}
\arguments{
\item{x}{}

\item{select}{}
}
\value{

}
\description{
Extractor for dlmFiltered objects
}
\author{
Søren Højsgaard
}
