% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIME.R
\name{LIME}
\alias{LIME}
\alias{lime}
\title{Length-based Integrated Mixed Effects (LIME) assessment model}
\usage{
LIME(
  x = 1,
  Data,
  add_catch = FALSE,
  add_index = FALSE,
  ESS = 50,
  CAL_dist = c("mult", "Dirmult"),
  SigmaC = 0.2,
  SigmaI = 0.2,
  SigmaR = Data@sigmaR[x],
  SigmaF = 0.3,
  nseas = 1L,
  Fproxy = c("F40", "F30", "Fmax"),
  yind = expression(1:ncol(Data@Cat)),
  integrate = FALSE,
  run_LIME_args,
  control = list(iter.max = 2e+05, eval.max = 4e+05),
  inner.control = list(),
  ...
)
}
\arguments{
\item{x}{An index for the objects in \code{Data} when running in \link[DLMtool]{runMSE}.
Otherwise, equals to 1 when running an assessment interactively.}

\item{Data}{An object of class Data.}

\item{add_catch}{Logical, whether to include catch data.}

\item{add_index}{Logical, whether to include index data.}

\item{ESS}{The maximum annual sample size of the length composition data.}

\item{CAL_dist}{Whether the model uses a multinomial \code{"mult"} or Dirichlet-multinomial \code{"Dirmult"}
likelihood for the length composition.}

\item{SigmaC}{The standard deviation of the catch in the likelihood.}

\item{SigmaI}{The standard deviation of the index in the likelihood.}

\item{SigmaR}{The standard deviation of the recruitment deviates in the likelihood.}

\item{SigmaF}{The standard deviation of the F random walk in the likelihood.}

\item{nseas}{The number of season in the model. In high F situations, multi-seasons may be needed to model smooth
length distributions.}

\item{Fproxy}{How the FMSY proxy reference point is calculated. The default is F40\%.}

\item{yind}{Optional, vector of years for the model. A subset of the data will be taken instead of the full
time series in the Data object.}

\item{integrate}{Logical, whether recruitment deviates are random effects in the model (TRUE) or penalized effects (FALSE).
FALSE by default.}

\item{run_LIME_args}{A named list of additional arguments for \link[LIME]{run_LIME}. Only arguments that are passed to \link[LIME]{format_input} are used.}

\item{control}{A named list of agruments for optimization to be passed to \code{\link[stats]{nlminb}}.}

\item{inner.control}{A named list of arguments for optimization of the random effects, which
is passed on to \code{\link[TMB]{newton}}.}

\item{...}{Additional arguments (not currently used).}
}
\value{
An object of \code{\linkS4class{Assessment}} containing assessment output.
}
\description{
A wrapper function for the LIME model, a stock assessment model designed for data-limited scenarios with
length composition data and biological parameters. Optionally, catch or index abundance data can
also be included. Catch data should be used in order to estimate stock biomass for output-based harvest
control rules. A steepness value of one is used in typical applications. The reported FMSY reference point is a
proxy using either spawning potential ratio or Fmax (from yield-per-recruit). The BMSY reference point is
calculated using estimated mean recruitment and fishing at FMSY.
}
\note{
This is a wrapper function intended for a DLMtool \linkS4class{Data} object. The \code{LIME} package can be
downloaded from Github with \code{devtools::install_github("merrillrudd/LIME")}.
}
\section{Required Data}{

\itemize{
\item \code{LIME}: Mort, L50, L95, CAL, CAL_bins, vbK, vbLinf, vbt0, wla, wlb, LenCV, sigmaR
}
}

\section{Optional Data}{

\itemize{
\item \code{LIME}: Cat, Ind
}
}

\examples{
\donttest{
library(MSEtool)
data(SimulatedData)

res <- LIME(Data = SimulatedData)
plot(res)
summary(res)

## Use additional LIME functions
output <- res@info
LIME::plot_LCfits(Inputs = output$Inputs, Report = output$Report)
LIME::plot_output(Inputs = output$Inputs, Report = output$Report, Sdreport = output$Sdreport)

## Create an MP that uses a F30\% proxy and catch data
LIME_MP <- make_MP(DLMextra::LIME, HCR40_10, Fproxy = "F30", add_catch = TRUE)
}
}
\references{
Rudd, M.B., and Thorson, J.T. 2017. Accounting for variable recruitment and fishing mortality in length-based
stock assessments for data-limited fisheries. Canadian Journal of Fisheries and Aquatic Sciences 75:1019-1035.
\url{https://doi.org/10.1139/cjfas-2017-0143}
}
\seealso{
\link[MSEtool]{plot.Assessment} \link[MSEtool]{summary.Assessment} \link[MSEtool]{make_MP}
}
\author{
Q. Huynh
}
