% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmXtra.R
\name{dlmXFilter}
\alias{dlmXFilter}
\title{Filter a dlm}
\usage{
dlmXFilter(object, debug = FALSE, simplify = FALSE)
}
\arguments{
\item{object}{Result from calling dlmXMLE or dlmXOpt}

\item{debug}{FALSE}

\item{simplify}{FALSE}
}
\value{
Same as when calling dlmFilter
}
\description{
A wrapper for dlmFilter

Not many details for now
}
\author{
Søren Højsgaard
}
