% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmXtra.R
\name{dlmXMLE}
\alias{dlmXMLE}
\title{Fit dlm with MLE}
\usage{
dlmXMLE(y, parm, build, method = "L-BFGS-B", ..., debug = FALSE)
}
\arguments{
\item{y}{Data}

\item{parm}{Intial value for parameter}

\item{build}{Build function}

\item{method}{Method used by optim}

\item{...}{Parameters passed on to build and optim}

\item{debug}{FALSE}
}
\value{
The result of calling optim padded with y and the fitted
    model object.
}
\description{
A wrapper for calling dlmMLE

Not many details for now
}
\author{
Søren Højsgaard
}
