% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmXtra.R
\name{extract-dlm}
\alias{extract-dlm}
\alias{getFF}
\alias{getGG}
\alias{getVV}
\alias{getWW}
\title{Extractor for dlm objects}
\usage{
getFF(mod, i)

getGG(mod, i)

getVV(mod, i)

getWW(mod, i)
}
\arguments{
\item{mod}{dlm object}

\item{i}{Row in X matrix}
}
\value{
The relevant result
}
\description{
Extractor for dlm objects
}
\author{
Søren Højsgaard
}
