% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extract-dlm}
\alias{extract-dlm}
\alias{getdlm}
\alias{getdlm.dlm}
\alias{getFF}
\alias{getGG}
\alias{getVV}
\alias{getWW}
\title{Extractor for dlm objects}
\usage{
getdlm(object, slot, i)

\method{getdlm}{dlm}(object, slot, i)

getFF(object, i)

getGG(object, i)

getVV(object, i)

getWW(object, i)
}
\arguments{
\item{object}{dlm object}

\item{slot}{Which slot in the object to select.}

\item{i}{Row in X matrix}
}
\value{
The relevant result
}
\description{
Extractor for dlm objects
}
\author{
Søren Højsgaard
}
