% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extract-dlmFiltered}
\alias{extract-dlmFiltered}
\alias{getdlmf}
\alias{getdlmf.dlmFiltered}
\title{Extractor for dlmFiltered objects}
\usage{
getdlmf(object, slot, select = TRUE)

\method{getdlmf}{dlmFiltered}(object, slot, select = TRUE)
}
\arguments{
\item{object}{A dlmFiltered object.}

\item{slot}{Which slot in the object to select.}

\item{select}{Which element to select.}
}
\value{
The relevant result
}
\description{
Extractor for dlmFiltered objects
}
\author{
Søren Højsgaard
}
