% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmxforecast.R
\name{forecast}
\alias{forecast}
\alias{dlmXForecast}
\alias{dlmXForecast.dlm}
\alias{dlmXForecast.dlmFiltered}
\title{Forecast for dlm and dlmFiltered objects}
\usage{
dlmXForecast(object, nAhead = 1, offset = 0)

\method{dlmXForecast}{dlm}(object, nAhead = 1, offset = 0)

\method{dlmXForecast}{dlmFiltered}(object, nAhead = 1, offset = 0)
}
\arguments{
\item{object}{dlm or dlmFiltered objects}

\item{nAhead}{Number of steps ahead}

\item{offset}{Where to start from}
}
\value{
A list with a, R, f, Q and other values
}
\description{
Forecast for dlm and dlmFiltered objects
}
\author{
Søren Højsgaard
}
